Fix all warnings

+++ src/graphics/bindless.cpp
@@ -21,17 +21,15 @@
 void BindlessRegistry::Init( GfxDevice &gfx ) {
     // Create descriptor pool
     {
-        VkDescriptorPoolSize pool_sizes[] = {
-                { VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, MaxBindlessImages },
-                { VK_DESCRIPTOR_TYPE_SAMPLER, MaxSamplers } };
+        VkDescriptorPoolSize pool_sizes[] = { { VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, MaxBindlessImages },
+                                              { VK_DESCRIPTOR_TYPE_SAMPLER, MaxSamplers } };
 
-        const VkDescriptorPoolCreateInfo info = {
-                .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO,
-                .pNext = nullptr,
-                .flags = VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT,
-                .maxSets = MaxBindlessImages * 2,
-                .poolSizeCount = 2u,
-                .pPoolSizes = pool_sizes };
+        const VkDescriptorPoolCreateInfo info = { .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO,
+                                                  .pNext = nullptr,
+                                                  .flags = VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT,
+                                                  .maxSets = MaxBindlessImages * 2,
+                                                  .poolSizeCount = 2u,
+                                                  .pPoolSizes = pool_sizes };
 
         VK_CHECK( vkCreateDescriptorPool( gfx.device, &info, nullptr, &pool ) );
     }
@@ -39,16 +37,14 @@ void BindlessRegistry::Init( GfxDevice &gfx ) {
     // Descriptor Set Layout
     {
         std::array bindings = {
-                VkDescriptorSetLayoutBinding{
-                        .binding = TextureBinding,
-                        .descriptorType = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-                        .descriptorCount = MaxBindlessImages,
-                        .stageFlags = VK_SHADER_STAGE_ALL },
-                VkDescriptorSetLayoutBinding{
-                        .binding = SamplersBinding,
-                        .descriptorType = VK_DESCRIPTOR_TYPE_SAMPLER,
-                        .descriptorCount = MaxSamplers,
-                        .stageFlags = VK_SHADER_STAGE_ALL },
+                VkDescriptorSetLayoutBinding{ .binding = TextureBinding,
+                                              .descriptorType = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
+                                              .descriptorCount = MaxBindlessImages,
+                                              .stageFlags = VK_SHADER_STAGE_ALL },
+                VkDescriptorSetLayoutBinding{ .binding = SamplersBinding,
+                                              .descriptorType = VK_DESCRIPTOR_TYPE_SAMPLER,
+                                              .descriptorCount = MaxSamplers,
+                                              .stageFlags = VK_SHADER_STAGE_ALL },
         };
 
         std::array<VkDescriptorBindingFlags, 2> binding_flags = {
@@ -58,7 +54,7 @@ void BindlessRegistry::Init( GfxDevice &gfx ) {
         VkDescriptorSetLayoutBindingFlagsCreateInfo flag_info = {
                 .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_BINDING_FLAGS_CREATE_INFO,
                 .pNext = nullptr,
-                .bindingCount = binding_flags.size( ),
+                .bindingCount = ( u32 )binding_flags.size( ),
                 .pBindingFlags = binding_flags.data( ),
         };
 
@@ -66,7 +62,7 @@ void BindlessRegistry::Init( GfxDevice &gfx ) {
                 .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO,
                 .pNext = &flag_info,
                 .flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT,
-                .bindingCount = bindings.size( ),
+                .bindingCount = ( u32 )bindings.size( ),
                 .pBindings = bindings.data( ) };
 
         VK_CHECK( vkCreateDescriptorSetLayout( gfx.device, &info, nullptr, &layout ) );
@@ -84,12 +80,11 @@ void BindlessRegistry::Init( GfxDevice &gfx ) {
         VK_CHECK( vkAllocateDescriptorSets( gfx.device, &alloc_info, &set ) );
 
 
-        const VkDebugUtilsObjectNameInfoEXT obj = {
-                .sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT,
-                .pNext = nullptr,
-                .objectType = VK_OBJECT_TYPE_DESCRIPTOR_SET,
-                .objectHandle = reinterpret_cast<uint64_t>( set ),
-                .pObjectName = "Bindless Descriptor Set" };
+        const VkDebugUtilsObjectNameInfoEXT obj = { .sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT,
+                                                    .pNext = nullptr,
+                                                    .objectType = VK_OBJECT_TYPE_DESCRIPTOR_SET,
+                                                    .objectHandle = reinterpret_cast<uint64_t>( set ),
+                                                    .pObjectName = "Bindless Descriptor Set" };
         vkSetDebugUtilsObjectNameEXT( gfx.device, &obj );
     }
 
@@ -153,16 +148,15 @@ void BindlessRegistry::InitSamplers( GfxDevice &gfx ) {
     }
 
     {
-        constexpr VkSamplerCreateInfo create_info = {
-                .sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO,
-                .magFilter = VK_FILTER_LINEAR,
-                .minFilter = VK_FILTER_LINEAR,
-                .mipmapMode = VK_SAMPLER_MIPMAP_MODE_LINEAR,
-                .mipLodBias = 0.0f,
-                .anisotropyEnable = VK_TRUE,
-                .maxAnisotropy = 16.0f,
-                .minLod = 0.0f,
-                .maxLod = 10.0f };
+        constexpr VkSamplerCreateInfo create_info = { .sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO,
+                                                      .magFilter = VK_FILTER_LINEAR,
+                                                      .minFilter = VK_FILTER_LINEAR,
+                                                      .mipmapMode = VK_SAMPLER_MIPMAP_MODE_LINEAR,
+                                                      .mipLodBias = 0.0f,
+                                                      .anisotropyEnable = VK_TRUE,
+                                                      .maxAnisotropy = 16.0f,
+                                                      .minLod = 0.0f,
+                                                      .maxLod = 10.0f };
         VK_CHECK( vkCreateSampler( gfx.device, &create_info, nullptr, &linearSampler ) );
         AddSampler( gfx, LinearSamplerId, linearSampler );
     }
